refactor(hooks): extract useGroupedStocks from bulk modals
