+++ src/components/modals/BulkSellModal.jsx
@@ -1,4 +1,5 @@
 import React, { useState, useMemo } from 'react';
+import { useGroupedStocks } from '../../hooks/useGroupedStocks';
 import { formatNumber, handleMKInput } from '../../utils/formatters';
 import '../../styles/bulk-modals.css';
 import StepInput from '../StepInput';
@@ -7,32 +8,7 @@ export default function BulkSellModal({ stocks, categories = [], tradeMode = 'tr
   const [searchQuery, setSearchQuery] = useState('');
   const [selectedItems, setSelectedItems] = useState({});
 
-  const groupedStocks = useMemo(() => {
-    const filtered = stocks.filter(s =>
-      (tradeMode === 'investment' ? s.isInvestment : !s.isInvestment) && s.shares > 0
-    );
-    const filteredCats = categories.filter(c =>
-      tradeMode === 'investment' ? c.isInvestment : !c.isInvestment
-    );
-    const catNames = filteredCats.map(c => c.name);
-
-    const groups = [];
-    for (const cat of filteredCats) {
-      const catStocks = filtered.filter(s => s.category === cat.name);
-      if (catStocks.length > 0) {
-        groups.push({ name: cat.name, stocks: catStocks });
-      }
-    }
-
-    const uncategorized = filtered.filter(s =>
-      s.category === 'Uncategorized' || !s.category || !catNames.includes(s.category)
-    );
-    if (uncategorized.length > 0 && !filteredCats.some(c => c.name === 'Uncategorized')) {
-      groups.push({ name: 'Uncategorized', stocks: uncategorized });
-    }
-
-    return groups;
-  }, [stocks, categories, tradeMode]);
+  const groupedStocks = useGroupedStocks(stocks, categories, tradeMode, { requireShares: true });
 
   const filteredGroups = useMemo(() => {
     if (!searchQuery.trim()) return groupedStocks;
